+++ activemodel/test/cases/type/unsigned_integer_test.rb
@@ -1,8 +1,9 @@
 require "cases/helper"
+require "active_model/type"
 
-module ActiveRecord
+module ActiveModel
   module Type
-    class UnsignedIntegerTest < ActiveRecord::TestCase
+    class UnsignedIntegerTest < ActiveModel::TestCase
       test "unsigned int max value is in range" do
         assert_equal(4294967295, UnsignedInteger.new.serialize(4294967295))
       end
